리뷰 좋아요는 배포 후 개인적으로 해보기로 함

+++ src/main/java/com/mangjakseon/dto/ReviewDTO.java
@@ -15,7 +15,7 @@ public class ReviewDTO {
 
     private String reviewContent;       //리뷰 내용
 
-    private Integer likeCount;         //좋아요 체크
+    private Long likeCount;         //좋아요 체크
 
     private double score;           //별점
